refactor: moved role middleware to authMiddleware
