+++ src/middleware/validationMiddleware.js
@@ -58,19 +58,3 @@ export const validateToken = async (req, res, next) => {
     }
   }
 };
-
-export const checkRole = (role) => {
-  return async (req, res, next) => {
-    const user = await User.findById(req.userId);
-    if (!user.roles.includes(role)) {
-      logger.warn('Unauthorized', {
-        method: req.method,
-        url: req.originalUrl,
-        userId: req.userId,
-        ip: req.ip,
-      });
-      return res.status(403).json({ message: 'Unauthorized' });
-    }
-    next();
-  };
-};
