+++ src/routes/userRoutes.js
@@ -1,9 +1,7 @@
 import express from 'express';
 import { createUser, login, logout } from '../controllers/userController.js';
-import {
-  validateToken,
-  checkRole,
-} from '../middleware/validationMiddleware.js';
+import { validateToken } from '../middleware/validationMiddleware.js';
+import { checkRole } from '../middleware/authMiddleware.js';
 
 const router = express.Router();
